update test runner

+++ test.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -x
 #
 # OpenTHC Test Runner
 #
@@ -11,45 +11,72 @@ d=$(dirname "$f")
 
 cd "$d"
 
-output_base="../webroot/test-output"
+output_base="webroot/test-output"
 output_main="$output_base/index.html"
 mkdir -p "$output_base"
 
-search_list="
-../boot.php
-../bin/
-../lib/
-../sbin/
-../test/
-../view/
-"
+code_list=(
+	boot.php
+	bin/
+	lib/
+	sbin/
+	test/
+	view/
+)
 
 
 #
 # Lint
-echo '<h1>Linting...</h1>' > "$output_main"
-find $search_list -type f -name '*.php' -exec php -l {} \; \
-	| grep -v 'No syntax' || true \
-	2>&1 >"$output_base/phplint.txt"
-[ -s "$output_base/phplint.txt" ] || echo "Linting OK" >"$output_base/phplint.txt"
+if [ ! -f "$output_base/phplint.txt" ]
+then
+
+	echo '<h1>Linting...</h1>' > "$output_main"
+
+	find "${code_list[@]}" -type f -name '*.php' -exec php -l {} \; \
+		| grep -v 'No syntax' || true \
+		2>&1 \
+		>"$output_base/phplint.txt"
+
+	[ -s "$output_base/phplint.txt" ] || echo "Linting OK" >"$output_base/phplint.txt"
+
+fi
+
+
+#
+# PHP-CPD
+if [ ! -f "$output_base/phpcpd.txt" ]
+then
+
+	echo '<h1>CPD Check</h1>' > "$output_main"
+
+	vendor/bin/phpcpd \
+		--fuzzy \
+		"${code_list[@]}" \
+		2>&1 \
+		> "$output_base/phpcpd.txt"
+
+fi
 
 
 #
 # PHPStan
-echo '<h1>PHPStan...</h1>' > "$output_main"
-../vendor/bin/phpstan analyze --error-format=junit --no-progress > "$output_base/phpstan.xml" || true
-[ -f "phpstan.xsl" ] || wget -q 'https://openthc.com/pub/phpstan.xsl'
-xsltproc \
-	--nomkdir \
-	--output "$output_base/phpstan.html" \
-	phpstan.xsl \
-	"$output_base/phpstan.xml"
+if [ ! -f "$output_base/phpstan.html" ]
+then
+	echo '<h1>PHPStan...</h1>' > "$output_main"
+	vendor/bin/phpstan analyze --error-format=junit --no-progress > "$output_base/phpstan.xml" || true
+	[ -f "phpstan.xsl" ] || curl -qs 'https://openthc.com/pub/phpstan.xsl' > "test/phpstan.xsl"
+	xsltproc \
+		--nomkdir \
+		--output "$output_base/phpstan.html" \
+		phpstan.xsl \
+		"$output_base/phpstan.xml"
+fi
 
 
 #
 # PHPUnit
 echo '<h1>PHPUnit...</h1>' > "$output_main"
-../vendor/bin/phpunit \
+vendor/bin/phpunit \
 	--verbose \
 	--log-junit "$output_base/phpunit.xml" \
 	--testdox-html "$output_base/testdox.html" \
@@ -83,6 +110,7 @@ cat <<HTML > "$output_main"
 <style>
 html {
 	font-family: sans-serif;
+	font-size: 1.5rem;
 }
 </style>
 <title>Test Result ${dt}</title>
